CELL moved to fninformation.go
